Add Cache-Control for Service Document Resource
We allow the service document to cache for 60 seconds. The E-Tag is
calculated from the media-type and the link hrefs because they can
change with context path. I use the pandect lib for MD5 hashing.

I had to create a nil returning not-modified handler because liberator
returns a text/plain response otherwise.