Remove Bang from Version Advancing Functions
Because the version is not changed by e.g. adding a query (a new version
is created), it makes no sense to have a bang on such functions. For
sure, there is some side effect, because a new version is created in
durable storage, but one can see this as simply creating garbage if
nobody needs the new version afterwards.